Api_backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
spotipy==2.22.1
psycopg2-binary==2.9.9
requests==2.31.0
PyJWT==2.8.0
APScheduler==3.10.4
